Fixed usage of yyFlexLexer with multiple parsers
