+++ frontends/parsers/p4/p4lexer.hpp
@@ -4,31 +4,12 @@
 // FlexLexer.h requires you to provide an external #include guard so that it can
 // be included multiple times, each time providing a different definition for
 // yyFlexLexer, to define multiple lexer base classes.
-#ifndef yyFlexLexerOnce
+
 #define yyFlexLexer p4FlexLexer
 #include <FlexLexer.h>
-#endif
+#undef yyFlexLexer
 
-#include "frontends/parsers/p4/abstractP4Lexer.hpp"
-#include "frontends/parsers/p4/p4parser.hpp"
-#include "lib/source_file.h"
+#include "frontends/parsers/p4/p4lexer_internal.hpp"
 
-namespace P4 {
-
-class P4ParserDriver;
-
-class P4Lexer : public AbstractP4Lexer, public p4FlexLexer {
- public:
-    explicit P4Lexer(std::istream& input)
-        : p4FlexLexer(&input), needStartToken(true) { }
-
-    virtual Token yylex(P4::P4ParserDriver& driver) override;
-
- private:
-    bool needStartToken;
-    int yylex() override { return p4FlexLexer::yylex(); }
-};
-
-}  // namespace P4
 
 #endif  /* FRONTENDS_P4_LEXER_H_ */
