+++ frontends/parsers/v1/v1parser.ypp
@@ -21,7 +21,7 @@ limitations under the License.
 // Set up names.
 %defines
 %define api.namespace {V1}
-%define parser_class_name {V1Parser}
+%define api.parser.class {V1Parser}
 
 // Use the C++-native variant representation of tokens.
 %define api.token.constructor
